requirements-py.txt
flask
pymongo
python-dotenv
flask_socketio